lots of refactoring with closure plot function - added closure_of_8_bit_systems.cpp - renamed and refactored draw_closure_plots.ipynb - Refactored generateClosurePlots.cpp to a header file to align with the libray the file is now called generateClosurePlots.hpp
